% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Amphibac.R
\docType{data}
\name{Amphibac}
\alias{Amphibac}
\title{Amphibac}
\format{
A .csv data frame with 1944 rows and 3 columns:
\describe{
  \item{variable1}{Description of variable1}
  \item{variable2}{Description of variable2}
  ...
}
}
\source{
\url{http://example.com/data-source}
}
\usage{
Amphibac
}
\description{
Amphibac database containing 16S rRNA sequences identified to possess anti-Bd potential.
}
\keyword{datasets}
