% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Amphibac.R
\docType{data}
\name{AmphiBac_2023.2_Facilitating}
\alias{AmphiBac_2023.2_Facilitating}
\title{AmphiBac_2023.2_Facilitating Dataset}
\format{
a data frame with 597 rows and 2 variables:
\describe{
  \item{species}{the species that the bacterial sequence was originally isolated from}
  \item{isolating_strain}{the name of the strain og the sequence}
  \item{ref_seq}{the sequence from the region of identification}
  ...
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
data(AmphiBac_2023.2_Facilitating)

a data frame with 597 rows and 2 variables
}
\description{
Dataframe containing facilitory sequences and the organism they were isolated from
}
\keyword{datasets}
