% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Amphibac.R
\name{AmphibacMatch_Facilitating}
\alias{AmphibacMatch_Facilitating}
\title{AmphibacMatch_Facilitating}
\usage{
AmphibacMatch_Facilitating(
  df,
  perc = 0,
  k = 0,
  score = 0,
  gap_start = 0,
  gap = 3,
  match = 1,
  mismatch = -3,
  ver = "2023"
)
}
\arguments{
\item{df}{A data frame containing 'ID' and 'Sequence' columns.}

\item{perc}{numerical percentage cutoff for minimum percent identity match in results df. Default is 0.}

\item{score}{a numerical cutoff for alignment score in results df. Default is 0.}

\item{gap_start}{The gap opening penalty. Default is 0.}

\item{gap}{The gap extension penalty. Default is 3.}

\item{match}{The score for a match. Default is 1.}

\item{mismatch}{The penalty for a mismatch. Default is -3.}

\item{ver}{The year of the database which the user desires to use. Default is the latest version of the database available}
}
\value{
A data frame with alignment results.
}
\description{
This function matches user-provided sequences against a reference database.
}
\examples{
user_df <- data.frame(ID = c("seq1", "seq2"), Sequence = c("ATCG", "AAGC"))
result <- AmphibacMatch(user_df)

}
