% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Amphibac.R
\name{AmphibacMatch_Strict_clustered}
\alias{AmphibacMatch_Strict_clustered}
\title{AmphibacMatch_Strict_clustered}
\usage{
AmphibacMatch_Strict_clustered(df, perc = 0, score = 0, ver = "2023")
}
\arguments{
\item{df}{A data frame containing 'ID' and 'Sequence' columns.}

\item{perc}{numerical percentage cutoff for minimum percent identity match in results df. Default is 0.}

\item{score}{a numerical cutoff for alignment score in results df. Default is 0.}

\item{gap_start}{The gap opening penalty. Default is 0.}

\item{gap}{The gap extension penalty. Default is 3.}

\item{match}{The score for a match. Default is 1.}

\item{mismatch}{The penalty for a mismatch. Default is -3.}

\item{database}{The name of the database to be used to compare df sequences against. Default is 'amphibac'.}
}
\value{
A data frame with alignment results.
}
\description{
This function matches user-provided sequences against a reference database.
}
\examples{
user_df <- data.frame(ID = c("seq1", "seq2"), Sequence = c("ATCG", "AAGC"))
result <- AmphibacMatch_Strict_clustered(user_df)

}
