% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Amphibac.R
\name{read.qza}
\alias{read.qza}
\title{read.qza}
\usage{
read.qza(file)
}
\arguments{
\item{qza}{The path to the FASTA file.}
}
\value{
A data frame from a QIIME2 derived relative frequency table
}
\description{
This function reads a .qza file derived from QIIME2 and delivers a editable dataframe in the r environment
}
