% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Amphibac.R
\name{ReadFasta}
\alias{ReadFasta}
\title{ReadFasta}
\usage{
ReadFasta(fasta)
}
\arguments{
\item{fasta}{The path to the FASTA file.}
}
\value{
A data frame containing species, isolating strain, and sequences.
}
\description{
This function reads a FASTA file and formats it for use with the Amphibac search function.
}
