% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Amphibac.R, R/data.R
\docType{data}
\name{Amphibac}
\alias{Amphibac}
\title{Amphibac Dataset}
\format{
a data frame with 1944 rows and 3 variables:
\describe{
  \item{species}{the species that the bacterial sequence was originally isolated from}
  \item{isolating_strain}{the name of the strain og the sequence}
  \item{ref_seq}{the sequence from the region of identification}
  ...
}

a data frame with 1944 rows and 3 variables:
\describe{
  \item{species}{the species that the bacterial sequence was originally isolated from}
  \item{isolating_strain}{the name of the strain og the sequence}
  \item{ref_seq}{the sequence from the region of identification}
  ...
}

a data frame with 1944 rows and 3 variables:
\describe{
  \item{species}{the species that the bacterial sequence was originally isolated from}
  \item{isolating_strain}{the name of the strain og the sequence}
  \item{ref_seq}{the sequence from the region of identification}
}
}
\source{
\url{http://somewhere.important.com/}

\url{http://somewhere.important.com/}

\url{http://somewhere.important.com/}
}
\usage{
data(Amphibac)

a data frame with 1944 rows and 3 variables

data(Amphibac)

a data frame with 1944 rows and 3 variables

Amphibac
}
\description{
Dataframe containing anti-batrachochytrium dendrobatidis sequences and the organism they were isolated from

Dataframe containing anti-batrachochytrium dendrobatidis sequences and the organism they were isolated from

Dataframe containing anti-batrachochytrium dendrobatidis sequences and the organism they were isolated from
}
\keyword{datasets}
